requirements.txt
streamlit>=1.32
pandas>=2.0
numpy>=1.26
faiss-cpu==1.7.4
sentence-transformers>=2.6
torch>=2.1,<3
openpyxl>=3.1
